[io] Export all channels by default; the first only if manually set.
This is a more reliable approach from the user perspective: by default,
uvplot extracts all the data available.
It would be bad if a user discovers that has been using only part of the
data due to this default.
For people that use to channel-average their spws into 1 channel per spw
this won't make any difference.